#7 - moved multiplayer changes into main track loader state
